// frontend/contracts/core/v1/models/ApproveOrganizationJoinRequestModel.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type ApproveOrganizationJoinRequestModel = {
    role?: ApproveOrganizationJoinRequestModel.role;
};
export namespace ApproveOrganizationJoinRequestModel {
    export enum role {
        OWNER = 'owner',
        ADMIN = 'admin',
        MEMBER = 'member',
    }
}