// frontend/contracts/core/v1/models/OrganizationInvitationModel.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type OrganizationInvitationModel = {
    id: string;
    organization_id: string;
    organization_name: string;
    organization_login: string;
    organization_avatar_url?: string;
    user_id: string;
    username: string;
    email?: string;
    inviter_id: string;
    inviter_username: string;
    role: string;
    status: string;
    created_at: string;
    updated_at: string;
};