// frontend/contracts/core/v1/models/OrganizationJoinRequestModel.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type OrganizationJoinRequestModel = {
    id: string;
    organization_id: string;
    organization_name: string;
    organization_login: string;
    user_id: string;
    username: string;
    email?: string;
    message?: string;
    status: string;
    reviewer_username?: string;
    created_at: string;
    updated_at: string;
};